incorporate changes suggested by @supereg
