+++ Sources/ApodiniExtension/Evaluation/EagerBuffer.swift
