+++ Sources/ApodiniExtension/RequestDecoding/DecodingStrategy/DecodingStrategy.swift
@@ -101,6 +101,3 @@ public extension DecodingStrategy {
         TransformingStrategy(self, using: transformer)
     }
 }
-
-
-// MARK: Implementations
