+++ Sources/ApodiniExtension/RequestDecoding/EndpointDecodingStrategy/EndpointDecodingStrategy.swift
@@ -149,7 +149,7 @@ private struct SomeEndpointDecodingStrategyCaller<S: EndpointDecodingStrategy>: 
 /// and a parent ``EndpointDecodingStrategy`` with ``EndpointDecodingStrategy/Input`` type `I`.
 ///
 /// This strategy allows for reuse of ``EndpointDecodingStrategy`` with a more basic ``EndpointDecodingStrategy/Input`` than
-/// the one you are acutally using. E.g. many of the predefined strategies have input type `Data`.
+/// the one you are actually using. E.g. many of the predefined strategies have input type `Data`.
 /// If your input type is a complete request, you can probably reuse those predefined strategies by extracting the body
 /// from the request.
 public struct TransformingEndpointStrategy<S: EndpointDecodingStrategy, I>: EndpointDecodingStrategy {
@@ -172,123 +172,5 @@ public extension EndpointDecodingStrategy {
     /// and a parent with input type `I`.
     func transformed<I>(_ transformer: @escaping (I) throws -> Self.Input) -> TransformingEndpointStrategy<Self, I> {
         TransformingEndpointStrategy(self, using: transformer)
-    }
-}
-
-
-// MARK: Implementations
-
-/// A ``EndpointDecodingStrategy`` that chooses between either of two strategies depending on the
-/// number of parameters on the given `endpoint` that are of Apodini `ParameterType/content`.
-public struct NumberOfContentParameterAwareStrategy<Input>: EndpointDecodingStrategy {
-    private let strategy: AnyEndpointDecodingStrategy<Input>
-    
-    /// Create a ``EndpointDecodingStrategy`` that chooses a different strategy depending on the
-    /// number of content parameters on the associated endpoint.
-    ///
-    /// - Parameters:
-    ///     - `endpoint`: The associated Apodini `Endpoint`
-    ///     - `one`: The strategy that is used if `endpoint` has one or zero parameters with `ParameterType/content`
-    ///     - `many`: The strategy that is used if `endpoint` has more than one parameter with `ParameterType/content`
-    public init<One: EndpointDecodingStrategy, Many: EndpointDecodingStrategy>(
-        for endpoint: AnyEndpoint,
-        using one: One,
-        or many: Many) where One.Input == Input, Many.Input == Input {
-        let onlyOneContentParameter = 1 <= endpoint[EndpointParameters.self].reduce(0, { count, parameter in
-            count + (parameter.parameterType == .content ? 1 : 0)
-        })
-                                            
-        if onlyOneContentParameter {
-            self.strategy = one.typeErased
-        } else {
-            self.strategy = many.typeErased
-        }
-    }
-    
-    public func strategy<Element>(for parameter: EndpointParameter<Element>)
-        -> AnyParameterDecodingStrategy<Element, Input> where Element: Decodable, Element: Encodable {
-        strategy.strategy(for: parameter)
-    }
-}
-
-public extension NumberOfContentParameterAwareStrategy where Input == Data {
-    /// An instance of ``NumberOfContentParameterAwareStrategy`` that chooses between
-    /// ``AllIdentityStrategy`` and ``AllNamedStrategy``.
-    static func oneIdentityOrAllNamedContentStrategy(_ decoder: AnyDecoder, for endpoint: AnyEndpoint) -> Self {
-        self.init(for: endpoint, using: AllIdentityStrategy(decoder), or: AllNamedStrategy(decoder))
-    }
-}
-
-/// An ``EndpointDecodingStrategy`` that uses a certain `AnyDecoder` to decode each
-/// parameter based on the parameters name.
-///
-/// The strategy expects the `Data` to hold a `String`-keyed container as its base element.
-/// The returned ``ParameterDecodingStrategy`` tries to decode the `Element` from
-/// the container that is keyed by `parameter/name`.
-///
-/// E.g. for `Element` being `String` and the parameter's name being `"name"`, the following JSON
-/// would be a valid input:
-///
-/// ```json
-/// {
-///     "name": "Max"
-/// }
-/// ```
-public struct AllNamedStrategy: EndpointDecodingStrategy {
-    private let decoder: AnyDecoder
-    
-    public init(_ decoder: AnyDecoder) {
-        self.decoder = decoder
-    }
-    
-    public func strategy<Element>(for parameter: EndpointParameter<Element>)
-        -> AnyParameterDecodingStrategy<Element, Data> where Element: Decodable, Element: Encodable {
-        NamedChildPatternStrategy<DynamicNamePattern<IdentityPattern<Element>>>(parameter.name, decoder).typeErased
-    }
-}
-
-/// An ``EndpointDecodingStrategy`` that chooses between two given strategies based
-/// on the `parameter`'s `parameterType`.
-public struct ParameterTypeSpecific<P: EndpointDecodingStrategy, B: EndpointDecodingStrategy>: EndpointDecodingStrategy where P.Input == B.Input {
-    private let backup: B
-    private let primary: P
-    private let parameterType: ParameterType
-    
-    /// Create an ``EndpointDecodingStrategy`` that uses different strategies based on the `parameter`'s
-    /// `parameterType`.
-    ///
-    /// - Parameters:
-    ///     - `type`: The Apodini `ParameterType` for which the `primary` strategy is used
-    ///     - `primary`: This strategy is used if the `parameter` is of `type`
-    ///     - `backup`: The strategy that is used if the `parameter` is not of `type`
-    public init(_ type: ParameterType = .content, using primary: P, otherwise backup: B) {
-        self.backup = backup
-        self.primary = primary
-        self.parameterType = type
-    }
-    
-    public func strategy<Element>(for parameter: EndpointParameter<Element>)
-        -> AnyParameterDecodingStrategy<Element, P.Input> where Element: Decodable, Element: Encodable {
-        if parameter.parameterType == self.parameterType {
-            return primary.strategy(for: parameter)
-        } else {
-            return backup.strategy(for: parameter)
-        }
-    }
-}
-
-/// An ``EndpointDecodingStrategy`` that uses a ``ParameterDecodingStrategy`` which
-/// refers to the given `exporter`'s ``LegacyInterfaceExporter/retrieveParameter(_:for:)``
-/// method to retrieve the `parameter`'s value.
-public struct InterfaceExporterLegacyStrategy<IE: LegacyInterfaceExporter>: EndpointDecodingStrategy {
-    private let exporter: IE
-    
-    public init(_ exporter: IE) {
-        self.exporter = exporter
-    }
-    
-    public func strategy<Element>(for parameter: EndpointParameter<Element>)
-        -> AnyParameterDecodingStrategy<Element, IE.ExporterRequest> where Element: Decodable, Element: Encodable {
-        InterfaceExporterLegacyParameterStrategy<IE, Element>(parameter: parameter, exporter: exporter).typeErased
     }
 }
